+++ CHANGELOG.md
@@ -134,3 +134,5 @@ darvin_content:
 - Remove unsupported translations.
 
 - Remove "forward-to-controller" widgets.
+
+- Rename SlugMapItem entity to ContentReference.